Fixed potential panic
